Move current dispatcher out of core
