+++ tokio-trace-core/src/dispatcher.rs
@@ -6,15 +6,9 @@ use {
 };
 
 use std::{
-    cell::RefCell,
     fmt,
     sync::{Arc, Weak},
-    thread,
 };
-
-thread_local! {
-    static CURRENT_DISPATCH: RefCell<Dispatch> = RefCell::new(Dispatch::none());
-}
 
 /// `Dispatch` trace data to a [`Subscriber`].
 #[derive(Clone)]
@@ -32,22 +26,6 @@ impl Dispatch {
         }
     }
 
-    pub(crate) fn with_current<T, F>(f: F) -> T
-    where
-        F: FnOnce(&Dispatch) -> T,
-    {
-        // If we try to access the current dispatcher while it's being
-        // dropped, `LocalKey::with` would panic, causing a double panic.
-        // However, we can't use `try_with` as we still need to invoke `f`,
-        // which would be captured by the closure.
-        if thread::panicking() {
-            // It's better to fail to collect instrumentation than cause a
-            // SIGSEGV.
-            return f(&Dispatch::none());
-        }
-        CURRENT_DISPATCH.with(|current| f(&*current.borrow()))
-    }
-
     /// Returns a `Dispatch` to the given [`Subscriber`](::Subscriber).
     pub fn new<S>(subscriber: S) -> Self
     // TODO: Add some kind of `UnsyncDispatch`?
@@ -59,28 +37,6 @@ impl Dispatch {
         };
         callsite::register_dispatch(&me);
         me
-    }
-
-    /// Sets this dispatch as the default for the duration of a closure.
-    ///
-    /// The default dispatcher is used when creating a new [`Span`] or
-    /// [`Event`], _if no span is currently executing_. If a span is currently
-    /// executing, new spans or events are dispatched to the subscriber that
-    /// tagged that span, instead.
-    ///
-    /// [`Span`]: ::span::Span
-    /// [`Subscriber`]: ::Subscriber
-    /// [`Event`]: ::Event
-    pub fn as_default<T>(&self, f: impl FnOnce() -> T) -> T {
-        if thread::panicking() {
-            return f();
-        }
-        CURRENT_DISPATCH.with(|current| {
-            let prior = current.replace(self.clone());
-            let result = f();
-            *current.borrow_mut() = prior;
-            result
-        })
     }
 
     pub(crate) fn registrar(&self) -> Registrar {
